Live Stream Tab
Implemented a live stream tab with a ifram and adjusted the local host with video and help, and then I was able to create the style kind of basic, but background, actually I'm going to take out the background. I changed that mid typing this and did it, so player container is gone and now is just the player. Time to test it, if not changed WE DID IT! See you there!